Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int num1 ,num2 , num3 , average ;
-    scanf("%d",&num1);
-    scanf("%d",&num2 );
-    scanf("%d",&num3);
+    float num1 ,num2 , num3 , average ;
+    scanf("%f",&num1);
+    scanf("%f",&num2 );
+    scanf("%f",&num3);
     average = (num1+num2+num3)/3;
     printf("Average: %.2f" , average);
